// mobile/types/character_types.ts
export type Religion = {
    god: string;
    devotion: number;
    [key: string]: any;
};